+++ CHANGELOG.md
@@ -7,3 +7,4 @@ All notable changes to `laravel-model-watch` will be documented in this file.
  - Initial version of model:watch
  - Monitor a single model
  - Customise the interval and fields
+ - Pass a ModelWatch Collection name for dynamically generated collections